Improve verification error handling
We should differentiate two cases:
- a verification error (e.g. decoding error, invalid signature, data not conforming to schema, kid not found in trust list, etc.)
- a general error occurred during verification (e.g. connection error when downloading a trust list, error accessing disk, out of memory, etc.)

In the first case, it does not make sense for the library client to retry the verification (for that particular hcert), since the hcert is most probably invalid.

In the second case, the problem most likely does not have to do with that hcert, and retrying could make sense. Also, in general, it could make sense to try to apply some corrective action, send an alert to an operator, etc.

Therefore there is the need to deal with both cases in different ways. This commit solves this problem by using a (new) VerificationException to propagate the errors belonging to the first use case.

Also a VerificationResult.errorMessage field has been added containing a more detailed message describing the error.