Return verification error details
For certain verification errors (initially, CWT_EXPIRED), return additional information in the form of key-value pairs.

The purpose of this is allowing the client code of the library to obtain additional data about certain errors, so that it can record richer logs or process that data in some way (e.g. presenting it to a human for diagnostics).

See issue #43 for a discussion of other alternatives to provide this feature.