JsDateSerializer: serialize the full date including time and tz info
JsDateSerializer is used for serializing the dates in the metaInformation
property of the validation result. Up until now it ignored any time
and timezone information which resulted in loss of information when
using the API from javascript.

Make it de/serialize the full ISO date string instead so no
information is lost.

Fixes #61