remove workaround for borked kotlin/js deps
Since https://youtrack.jetbrains.com/issue/KT-43237
is fixed (also for all our dependencies), we can drop the
workaround.

This also means that we could probably add a second JS target
without all the browserified node dependencies and directly
publish to npmjs.com in the future.
Hence, this paves the way to addressing #49.